Update navbar to handle the buttonConnect
